Move shared API type into SDK package
