+++ pkg/backend/httpstate/client/client.go
@@ -1370,26 +1370,7 @@ func (pc *Client) SubmitAIPrompt(ctx context.Context, requestBody interface{}) (
 	return res, err
 }
 
-type PublishPackageInput struct {
-	// Source is the source of the package. Typically this is 'pulumi' for packages published to the Pulumi Registry.
-	// Packages loaded from other registries (e.g. 'opentofu') will point to the origin of the package.
-	Source string
-	// Publisher is the organization that is publishing the package.
-	Publisher string
-	// Name is the URL-safe name of the package.
-	Name string
-	// Version is the semantic version of the package that should get published.
-	Version semver.Version
-	// Schema is a reader containing the JSON schema of the package.
-	Schema io.Reader
-	// Readme is a reader containing the markdown content of the package's README.
-	Readme io.Reader
-	// InstallDocs is a reader containing the markdown content of the package's installation documentation.
-	// This is optional, and if omitted, the package will not have installation documentation.
-	InstallDocs io.Reader
-}
-
-func (pc *Client) PublishPackage(ctx context.Context, input PublishPackageInput) error {
+func (pc *Client) PublishPackage(ctx context.Context, input apitype.PackagePublishOp) error {
 	req := apitype.StartPackagePublishRequest{
 		Version: input.Version.String(),
 	}
