+++ pkg/cmd/pulumi/packagecmd/package_publish.go
@@ -28,6 +28,7 @@ import (
 	cmdBackend "github.com/pulumi/pulumi/pkg/v3/cmd/pulumi/backend"
 	"github.com/pulumi/pulumi/pkg/v3/codegen/schema"
 	pkgWorkspace "github.com/pulumi/pulumi/pkg/v3/workspace"
+	"github.com/pulumi/pulumi/sdk/v3/go/common/apitype"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/env"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/resource/plugin"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/cmdutil"
@@ -204,7 +205,7 @@ func (cmd *packagePublishCmd) Run(
 		return fmt.Errorf("failed to read readme file: %w", err)
 	}
 
-	publishInput := backend.PackagePublishOp{
+	publishInput := apitype.PackagePublishOp{
 		Source:    args.source,
 		Publisher: publisher,
 		Name:      name,
